Fix bug with overwriting of job root file with each syst
I've just renamed the output files to be unique. In its docs, uproot says it supports root file insertion via the uproot.update function (as opposed to .create or .recreate) but it is broken.

In uproot/write/TFile.py as of uproot==3.10.12; uproot-methods==0.7.1

  27 class TFileUpdate(object):
  28     def __init__(self, path):
  29         self._openfile(path)
  30         raise NotImplementedError
  31
  32     def _openfile(self, path, compression):

As you can see _openfile is called with improper arguments in __init__ and there is even a raise NotImplementedError which is kind of a disclaimer about this class.

Hopefully we can get this fixed with the uproot guys, or ship a custom TFile.py...